Add initial commit for Vale compatible style guide
Two rules, a test file, some README's, added. Seperate *styles* will be
added for different rule sets, such as for the blog and Confluence will
be implemented later. The main README.md is very sparse, and will be
expanded as well. A PR Template will be added later as well.

QA: Using `_vale.ini` and `Fio-docs`, the repository markdown files was
checked, and purposeful errors then removed.

This commit applies to Jira FFTK-1624 and Jira FFTK-1627

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>