Add Branding style rule
While more will need to be identified and added, some of the common
branding and product name errors have been added. The use of regex can
be improved in the future as well. It is important to note that matches
for the expected form are (mostly) excluded.

The testing file was also added to, however it should be dramatically
cleaned up and structured to be a bit more systematic...currently it is
akin to scratch paper.

QA Steps: I ran vale against `test/test.md` as well as well as some PR
reviews.

This applies to Jira FFTK-1628

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>