Add Vocab accept file
A list of terms to accept have been added. This can be used across vale
styles, useful when working with styles other than fio-style.

Note there are times that we may not want to use the vocab, especially
while developing and fine tuning our own style.

This was used while running vale against docs. It is not case sensitive
at the moment, which we will likely want to address later.

This applies to Jira FFTK-1628

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>