Add rule for no click-here links
Added. This was done with scope set to link. It may not work in all
cases. Set to error as there's no reason to use "click here", and is a
easy fix. It matches links with "here" as "click" may not appear in the
link, or may be a variant such as "see here".

Tested this by linting some local files, but kept the tests fairly
simple and did not test on all markup languages.

This addresses Jira FFTK 1974

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>