Add rules for dashes
Rules for Em dashes and en dashes added, with one file for each. The em
dash one may prove to be a little too aggressive in matching, and is
fairly broad in scope, which we may want to change later. As such, the
em dash rule is set as a warning. The en dash should hopefully only be
triggered when appropriate. Note the en dash rule currently only covers
numerical ranges and dates. The use of en dashes extend past this,
however their use between words is more nuanced and will remain
something for a copy editor or reviewer to look out for.

QA steps: ran both against a set of misuses written in markdown, with
tests meant to trigger and not trigger the linter. The real test will be
how these punctuation marks end up being used or misused in content.

This commit addresses Jira FSTK #1980
This commit applies to FSTK #1628

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>