Add rule for Oxford comma
Added rule for oxford comma, setting it to warning. In most cases that
this matches, we will want to use it, however there are instances where
it is correct to avoid, such as with grouping of terms with a
conjunction between two terms.

QA Steps: Ran linter rule against each case, triggered only when
appropriate.

This commit addresses Jira FFTK #1982
This commit applies to Jira FFTK #1628

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>