Add trademark linter rules
Small set of trademarks added. Registered trademarks are caught, however
"TM" is not being caught, however this may be an encoding issue. Leaving
in for the time being, and this will be noted and worked on separately.

For QA steps, basic tests of terms written in markdown added, testing for multiple
instances, and across page, section, and paragraph. Working as intended.

This addresses Jira  FFTK #1978
This applies to Jira FFTK #1628

Signed-off-by: Katrina Prosise <katrina.prosise@foundries.io>