Fixup: em-dash rule
A style guide tells to use an em dash as an emphasis inside a sentence.
A corresponding rule should deny the use of hyphen or en dash in these places.
But, a rule in question was errorneously denying the use of hyphen or em dash.

This little fix was verified when running vale on a text using em dash.

Signed-off-by: Volodymyr Khoroz <volodymyr.khoroz@foundries.io>